🐞 Bugs in Samples 🐞 updates
