+++ Bugs in Samples/Bugs in Samples.csproj
@@ -13,20 +13,18 @@
   </ItemGroup>
 
   <ItemGroup>
-    <None Update="Shared/MasterView.razor.css" CssScope="master-view-scope" />
-    <None Update="Pages/MasterView/Buttons.razor.css" CssScope="buttons-scope" />
-    <None Update="Pages/MasterView/ButtonGroup.razor.css" CssScope="button-group-scope" />
-    <None Update="Pages/MasterView/ButtonStyles.razor.css" CssScope="button-styles-scope" />
-    <None Update="Pages/MasterView/CardActions.razor.css" CssScope="card-actions-scope" />
-    <None Update="Pages/MasterView/CardsOpacity.razor.css" CssScope="cards-opacity-scope" />
-    <None Update="Pages/MasterView/Chips.razor.css" CssScope="chips-scope" />
-    <None Update="Pages/MasterView/Dialog.razor.css" CssScope="dialog-scope" />
-    <None Update="Pages/MasterView/GridinTabLayout.razor.css" CssScope="gridin-tab-layout-scope" />
-    <None Update="Pages/MasterView/InputsCombos.razor.css" CssScope="inputs-combos-scope" />
-    <None Update="Pages/MasterView/NavigationBar.razor.css" CssScope="navigation-bar-scope" />
-    <None Update="Pages/MasterView/SpaceBetween.razor.css" CssScope="space-between-scope" />
-    <None Update="Shared/MasterView2.razor.css" CssScope="master-view-2-scope" />
-    <None Update="Pages/MasterView2/ResizingIssue.razor.css" CssScope="resizing-issue-scope" />
+    <None Update="Shared/MainLayout.razor.css" CssScope="master-view-scope" />
+    <None Update="Pages/Buttons.razor.css" CssScope="buttons-scope" />
+    <None Update="Pages/ButtonGroup.razor.css" CssScope="button-group-scope" />
+    <None Update="Pages/ButtonStyles.razor.css" CssScope="button-styles-scope" />
+    <None Update="Pages/CardActions.razor.css" CssScope="card-actions-scope" />
+    <None Update="Pages/CardsOpacity.razor.css" CssScope="cards-opacity-scope" />
+    <None Update="Pages/Chips.razor.css" CssScope="chips-scope" />
+    <None Update="Pages/Dialog.razor.css" CssScope="dialog-scope" />
+    <None Update="Pages/GridinTabLayout.razor.css" CssScope="gridin-tab-layout-scope" />
+    <None Update="Pages/InputsCombos.razor.css" CssScope="inputs-combos-scope" />
+    <None Update="Pages/NavigationBar.razor.css" CssScope="navigation-bar-scope" />
+    <None Update="Pages/SpaceBetween.razor.css" CssScope="space-between-scope" />
   </ItemGroup>
 
 </Project>
