+++ Bugs in Samples/Pages/ButtonGroup.razor.css
