+++ Bugs in Samples/Pages/ButtonStyles.razor
