+++ Bugs in Samples/Pages/ButtonStyles.razor.css
