+++ Bugs in Samples/Pages/Buttons.razor
