+++ Bugs in Samples/Pages/CardActions.razor.css
