+++ Bugs in Samples/Pages/CardsOpacity.razor
