+++ Bugs in Samples/Pages/Chips.razor
