+++ Bugs in Samples/Pages/Dialog.razor
