+++ Bugs in Samples/Pages/GridinTabLayout.razor
