+++ Bugs in Samples/Pages/InputsCombos.razor
