--- Bugs in Samples/Pages/MasterView/NavigationBar.razor
@@ -1,98 +1,0 @@
-@page "/navigation-bar"
-
-<div class="column-layout navigation-bar-container">
-    <div class="column-layout group">
-        <IgbNavbar class="navbar" navigation-bar-scope>
-            <h6>I only added the avatar to this one</h6>
-            <div slot="start">
-                <IgbIconButton Variant="IconButtonVariant.Flat" class="icon-button" navigation-bar-scope>
-                    <span class="material-icons">
-                        menu
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </IgbIconButton>
-            </div>
-            <div slot="end">
-                <IgbIconButton Variant="IconButtonVariant.Flat" class="icon-button" navigation-bar-scope>
-                    <span class="material-icons">
-                        search
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </IgbIconButton>
-                <IgbIconButton Variant="IconButtonVariant.Flat" class="icon-button" navigation-bar-scope>
-                    <span class="material-icons">
-                        favorite
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </IgbIconButton>
-                <IgbIconButton Variant="IconButtonVariant.Flat" class="icon-button" navigation-bar-scope>
-                    <span class="material-icons">
-                        more_vert
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </IgbIconButton>
-                <IgbAvatar Src="assets/Avatar2.png" Shape="AvatarShape.Circle" class="avatar" navigation-bar-scope></IgbAvatar>
-            </div>
-        </IgbNavbar>
-        <IgbNavbar class="navbar_1" navigation-bar-scope>
-            <h6>Rapid Fission Inc.</h6>
-            <IgbButton class="button" navigation-bar-scope>
-                HOME
-                <IgbRipple></IgbRipple>
-            </IgbButton>
-            <IgbButton Variant="ButtonVariant.Flat" class="button" navigation-bar-scope>
-                MY INFO
-                <IgbRipple></IgbRipple>
-            </IgbButton>
-            <IgbButton Variant="ButtonVariant.Flat" Disabled="true" class="button" navigation-bar-scope>
-                PEople
-                <IgbRipple></IgbRipple>
-            </IgbButton>
-            <IgbButton Variant="ButtonVariant.Flat" Disabled="true" class="button" navigation-bar-scope>
-                MY FILES
-                <IgbRipple></IgbRipple>
-            </IgbButton>
-            <div slot="end">
-                <div class="row-layout group_1"></div>
-                <IgbIconButton Variant="IconButtonVariant.Flat">
-                    <span class="material-icons">
-                        notifications
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </IgbIconButton>
-                <IgbAvatar Src="assets/Avatar6.png" Shape="AvatarShape.Circle" class="avatar" navigation-bar-scope></IgbAvatar>
-            </div>
-        </IgbNavbar>
-        <IgbNavbar class="navbar_1" navigation-bar-scope>
-            <h6>Rapid Fission Inc.</h6>
-            <IgbButton class="button" navigation-bar-scope>
-                HOME
-                <IgbRipple></IgbRipple>
-            </IgbButton>
-            <IgbButton Variant="ButtonVariant.Flat" class="button" navigation-bar-scope>
-                MY INFO
-                <IgbRipple></IgbRipple>
-            </IgbButton>
-            <IgbButton Variant="ButtonVariant.Flat" Disabled="true" class="button" navigation-bar-scope>
-                PEople
-                <IgbRipple></IgbRipple>
-            </IgbButton>
-            <IgbButton Variant="ButtonVariant.Flat" Disabled="true" class="button" navigation-bar-scope>
-                MY FILES
-                <IgbRipple></IgbRipple>
-            </IgbButton>
-            <div slot="end">
-                <div class="row-layout group_1"></div>
-                <IgbIconButton Variant="IconButtonVariant.Flat">
-                    <span class="material-icons">
-                        notifications
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </IgbIconButton>
-            </div>
-        </IgbNavbar>
-    </div>
-</div>
-
-@code {
-}
