--- Bugs in Samples/Pages/MasterView/_Imports.razor
@@ -1,1 +1,0 @@
-@layout MasterView
