--- Bugs in Samples/Pages/MasterView2/ResizingIssue.razor.css
@@ -1,194 +1,0 @@
-.row-layout {
-    display: flex;
-}
-.resizing-issue-container {
-    justify-content: flex-start;
-    align-items: stretch;
-    align-content: flex-start;
-    height: 100%;
-}
-.column-layout {
-    display: flex;
-    flex-direction: column;
-}
-.left_nav {
-    background-color: hsla(var(--ig-surface-500));
-    justify-content: flex-start;
-    align-items: stretch;
-    align-content: flex-start;
-    gap: 24px;
-    overflow: auto;
-    padding: 24px;
-    width: 260px;
-    flex-shrink: 0;
-}
-.group {
-    justify-content: flex-start;
-    align-items: center;
-    align-content: flex-start;
-    position: relative;
-    padding: 0 16px 0 0;
-    min-width: 50px;
-}
-.group_1 {
-    justify-content: flex-start;
-    align-items: stretch;
-    align-content: flex-start;
-    gap: 16px;
-    position: relative;
-    height: max-content;
-    min-width: 50px;
-    min-height: 50px;
-    flex-shrink: 0;
-}
-.right-content-1 {
-    background-color: rgba(96, 96, 96, 0.1);
-    justify-content: flex-start;
-    align-items: stretch;
-    align-content: flex-start;
-    gap: 24px;
-    overflow: auto;
-    position: relative;
-    padding: 24px;
-    min-width: 50px;
-    min-height: 50px;
-    flex-grow: 1;
-    flex-shrink: 0;
-}
-.content_toolbar-1 {
-    justify-content: flex-start;
-    align-items: center;
-    align-content: flex-start;
-    position: relative;
-    min-width: max-content;
-}
-.group_2 {
-    justify-content: flex-start;
-    align-items: stretch;
-    align-content: flex-start;
-    gap: 8px;
-    position: relative;
-    min-width: 50px;
-}
-.group_3 {
-    justify-content: flex-start;
-    align-items: stretch;
-    align-content: flex-start;
-    gap: 32px;
-    position: relative;
-    min-width: 50px;
-    min-height: 50px;
-    flex-grow: 1;
-    flex-shrink: 0;
-}
-.group_4 {
-    justify-content: flex-start;
-    align-items: stretch;
-    align-content: flex-start;
-    gap: 32px;
-    overflow: hidden;
-    position: relative;
-    min-width: 375px;
-    min-height: 50px;
-    flex-grow: 1;
-    flex-shrink: 0;
-}
-.group_5 {
-    justify-content: flex-start;
-    align-items: stretch;
-    align-content: flex-start;
-    gap: 16px;
-    position: relative;
-    min-width: 50px;
-    min-height: 50px;
-}
-.group_6 {
-    justify-content: flex-start;
-    align-items: stretch;
-    align-content: flex-start;
-    position: relative;
-    min-width: 50px;
-}
-.group_7 {
-    justify-content: flex-start;
-    align-items: stretch;
-    align-content: flex-start;
-    position: relative;
-    min-width: 50px;
-    min-height: 50px;
-}
-.group_8 {
-    justify-content: flex-start;
-    align-items: stretch;
-    align-content: flex-start;
-    gap: 32px;
-    overflow: hidden;
-    position: relative;
-    min-width: 50px;
-    min-height: 50px;
-    flex-shrink: 0;
-}
-.button {
-    left: 50%;
-    transform: translateX(-50%);
-    top: 40px;
-    position: absolute;
-    width: max-content;
-    height: max-content;
-}
-.button_1 {
-    width: max-content;
-    height: max-content;
-    min-width: min-content;
-}
-.button_2 {
-    height: max-content;
-    min-width: min-content;
-    flex-shrink: 0;
-}
-.content {
-    height: max-content;
-    min-width: min-content;
-}
-.content_1 {
-    height: max-content;
-    min-width: min-content;
-    flex-grow: 1;
-    flex-shrink: 0;
-}
-.icon {
-    font-size: 18px;
-    width: 18px;
-    height: 18px;
-    color: hsla(var(--ig-gray-600));
-}
-.avatar {
-    --background: none;
-}
-.calendar {
-    width: max-content;
-    height: max-content;
-}
-.icon_1 {
-    font-size: 24px;
-    width: 24px;
-    height: 24px;
-    color: #E83180;
-}
-.icon_2 {
-    font-size: 24px;
-    width: 24px;
-    height: 24px;
-    color: #72DA67;
-}
-.icon-button::part(base) {
-    color: hsla(var(--ig-gray-600));
-}
-.badge {
-    width: max-content;
-    height: max-content;
-}
-.list {
-    height: max-content;
-    min-width: min-content;
-}
