--- Bugs in Samples/Pages/MasterView2/_Imports.razor
@@ -1,1 +1,0 @@
-@layout MasterView2
