+++ Bugs in Samples/Pages/NavigationBar.razor.css
@@ -19,16 +19,6 @@
     min-width: 50px;
     min-height: 50px;
 }
-.row-layout {
-    display: flex;
-}
-.group_1 {
-    justify-content: flex-start;
-    align-items: center;
-    align-content: flex-start;
-    width: max-content;
-    height: max-content;
-}
 .navbar {
     background-color: hsla(var(--ig-secondary-700));
     color: hsla(var(--ig-surface-500));
@@ -37,7 +27,9 @@
 }
 .navbar_1 {
     background-color: hsla(var(--ig-surface-500));
+    color: hsla(var(--ig-gray-800));
     height: max-content;
+    min-width: min-content;
 }
 .icon-button::part(base) {
     color: hsla(var(--ig-surface-500));
@@ -49,6 +41,35 @@
     height: max-content;
     min-width: min-content;
 }
-.button::part(base) {
+.button_1::part(base) {
+    color: hsla(var(--ig-secondary-900));
+    background-color: hsla(var(--ig-surface-500));
+}
+.button_2::part(base) {
+    color: hsla(var(--ig-surface-500));
+}
+.button_3::part(base) {
+    color: hsla(var(--ig-surface-500));
+}
+.button_4::part(base) {
+    color: hsla(var(--ig-surface-500));
+}
+.button_5::part(base) {
+    color: hsla(var(--ig-secondary-900));
+    background-color: hsla(var(--ig-surface-500));
+}
+.button_6::part(base) {
+    color: hsla(var(--ig-surface-500));
+}
+.button_7::part(base) {
+    color: hsla(var(--ig-surface-500));
+}
+.button_8::part(base) {
+    color: hsla(var(--ig-surface-500));
+}
+.button_9::part(base) {
+    color: hsla(var(--ig-surface-500));
+}
+.button_10::part(base) {
     color: hsla(var(--ig-surface-500));
 }
