+++ Bugs in Samples/Program.cs
@@ -31,8 +31,6 @@ void RegisterIgniteUI(IServiceCollection services)
         typeof(IgbChipModule),
         typeof(IgbGridModule),
         typeof(IgbTabsModule),
-        typeof(IgbInputModule),
-        typeof(IgbDropdownHeaderModule),
-        typeof(IgbCalendarModule)
+        typeof(IgbInputModule)
     );
 }
