--- Bugs in Samples/Shared/MasterView.razor
@@ -1,141 +1,0 @@
-@inherits LayoutComponentBase
-@inject NavigationManager NavigationManager
-
-<div class="column-layout master-view-container">
-    <IgbNavbar class="navbar" master-view-scope>
-        <div slot="start">
-            <IgbIconButton Variant="IconButtonVariant.Flat" @onclick="@(e => _navDrawer!.ToggleAsync())" class="icon-button" master-view-scope>
-                <span class="material-icons">
-                    menu
-                </span>
-                <IgbRipple></IgbRipple>
-            </IgbIconButton>
-        </div>
-        <p class="typography__subtitle-2 text">
-            BUGS IN SAMPLES
-        </p>
-    </IgbNavbar>
-    <div class="row-layout group">
-        <IgbNavDrawer Open="true" Position="NavDrawerPosition.Relative" @ref="_navDrawer" class="nav-drawer" master-view-scope>
-            <IgbNavDrawerItem @onclick="@(e => Navigate(""))">
-                <span slot="icon">
-                    <span class="material-icons icon">
-                        bug_report
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </span>
-                <div slot="content">Buttons</div>
-            </IgbNavDrawerItem>
-            <IgbNavDrawerItem @onclick="@(e => Navigate("button-group"))">
-                <span slot="icon">
-                    <span class="material-icons icon">
-                        bug_report
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </span>
-                <div slot="content">Button Group</div>
-            </IgbNavDrawerItem>
-            <IgbNavDrawerItem @onclick="@(e => Navigate("button-styles"))">
-                <span slot="icon">
-                    <span class="material-icons icon">
-                        bug_report
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </span>
-                <div slot="content">Button Styles</div>
-            </IgbNavDrawerItem>
-            <IgbNavDrawerItem @onclick="@(e => Navigate("card-actions"))">
-                <span slot="icon">
-                    <span class="material-icons icon">
-                        bug_report
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </span>
-                <div slot="content">Card Action Buttons</div>
-            </IgbNavDrawerItem>
-            <IgbNavDrawerItem @onclick="@(e => Navigate("cards-opacity"))">
-                <span slot="icon">
-                    <span class="material-icons icon">
-                        bug_report
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </span>
-                <div slot="content">Cards Opacity</div>
-            </IgbNavDrawerItem>
-            <IgbNavDrawerItem @onclick="@(e => Navigate("chips"))">
-                <span slot="icon">
-                    <span class="material-icons icon">
-                        bug_report
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </span>
-                <div slot="content">Chips</div>
-            </IgbNavDrawerItem>
-            <IgbNavDrawerItem @onclick="@(e => Navigate("dialog"))">
-                <span slot="icon">
-                    <span class="material-icons icon">
-                        bug_report
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </span>
-                <div slot="content">Dialogs</div>
-            </IgbNavDrawerItem>
-            <IgbNavDrawerItem @onclick="@(e => Navigate("gridin-tab-layout"))">
-                <span slot="icon">
-                    <span class="material-icons icon">
-                        bug_report
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </span>
-                <div slot="content">Grid in Tab Layout</div>
-            </IgbNavDrawerItem>
-            <IgbNavDrawerItem @onclick="@(e => Navigate("inputs-combos"))">
-                <span slot="icon">
-                    <span class="material-icons icon">
-                        bug_report
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </span>
-                <div slot="content">Inputs &amp; Combos</div>
-            </IgbNavDrawerItem>
-            <IgbNavDrawerItem @onclick="@(e => Navigate("master-view-2"))">
-                <span slot="icon">
-                    <span class="material-icons icon">
-                        bug_report
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </span>
-                <div slot="content">NavBar</div>
-            </IgbNavDrawerItem>
-            <IgbNavDrawerItem @onclick="@(e => Navigate("master-view-2"))">
-                <span slot="icon">
-                    <span class="material-icons icon">
-                        bug_report
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </span>
-                <div slot="content">Resizing Issue</div>
-            </IgbNavDrawerItem>
-            <IgbNavDrawerItem @onclick="@(e => Navigate("space-between"))">
-                <span slot="icon">
-                    <span class="material-icons icon">
-                        bug_report
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </span>
-                <div slot="content">Space Between</div>
-            </IgbNavDrawerItem>
-        </IgbNavDrawer>
-        <div class="view-container">
-            @Body
-        </div>
-    </div>
-</div>
-
-@code {
-    private IgbNavDrawer? _navDrawer;
-    private void Navigate(string path)
-    {
-        NavigationManager.NavigateTo(path);
-    }
-}
