--- Bugs in Samples/Shared/MasterView.razor.css
@@ -1,57 +1,0 @@
-.column-layout {
-    display: flex;
-    flex-direction: column;
-}
-.master-view-container {
-    justify-content: flex-start;
-    align-items: stretch;
-    align-content: flex-start;
-    height: 100%;
-}
-.row-layout {
-    display: flex;
-}
-.group {
-    justify-content: flex-start;
-    align-items: stretch;
-    align-content: flex-start;
-    position: relative;
-    min-width: 50px;
-    min-height: 50px;
-    flex-grow: 1;
-    flex-basis: 0;
-}
-.navbar {
-    background-color: hsla(var(--ig-surface-500));
-    color: hsla(var(--ig-gray-800));
-    height: max-content;
-    min-width: min-content;
-}
-.nav-drawer {
-    min-width: min-content;
-    min-height: 0;
-    flex-shrink: 0;
-}
-.view-container {
-    overflow: auto;
-    position: relative;
-    min-width: 0;
-    min-height: 0;
-    flex-grow: 1;
-}
-.nav-drawer::part(main) {
-    width: 240px;
-}
-.icon-button::part(base) {
-    color: hsla(var(--ig-gray-700));
-}
-.text {
-    height: max-content;
-    min-width: min-content;
-}
-.icon {
-    font-size: 24px;
-    width: 24px;
-    height: 24px;
-    color: hsla(var(--ig-secondary-600));
-}
