--- Bugs in Samples/Shared/MasterView2.razor
@@ -1,143 +1,0 @@
-@inherits LayoutComponentBase
-
-<div class="column-layout master-view-2-container">
-    <div class="row-layout header">
-        <div class="row-layout group">
-            <img src="assets/slingshot-icon.svg" class="image" />
-            <IgbButton Variant="ButtonVariant.Flat" @onclick="@(e => _projectDropdown!.ToggleAsync(_button3))" @ref="_button3" class="button" master-view-2-scope>
-                <span>Home</span>
-                <span class="material-icons">
-                    keyboard_arrow_down
-                </span>
-                <IgbRipple></IgbRipple>
-            </IgbButton>
-            <IgbDropdown @ref="_projectDropdown" class="project-dropdown" master-view-2-scope>
-                <IgbDropdownItem>
-                    Home
-                </IgbDropdownItem>
-                <IgbDropdownItem>
-                    Company Inc.
-                </IgbDropdownItem>
-                <IgbDropdownHeader>
-                    INDIGO.DESIGN
-                </IgbDropdownHeader>
-                <IgbDropdownItem>
-                    Growth hacking
-                </IgbDropdownItem>
-                <IgbDropdownItem>
-                    Public preview
-                </IgbDropdownItem>
-                <IgbDropdownItem>
-                    UX Feedback
-                </IgbDropdownItem>
-                <IgbDropdownHeader>
-                    Reveal
-                </IgbDropdownHeader>
-                <IgbDropdownItem>
-                    Onboarding
-                </IgbDropdownItem>
-                <IgbDropdownItem>
-                    Analytics
-                </IgbDropdownItem>
-            </IgbDropdown>
-        </div>
-        <div class="row-layout group_1">
-            <IgbButton class="button_1" master-view-2-scope>
-                <span class="material-icons">
-                    add
-                </span>
-                <span>Task</span>
-                <IgbRipple></IgbRipple>
-            </IgbButton>
-            <IgbIconButton Variant="IconButtonVariant.Flat" class="icon-button" master-view-2-scope>
-                <span class="material-icons">
-                    add_alert
-                </span>
-                <IgbRipple></IgbRipple>
-            </IgbIconButton>
-            <IgbIconButton Variant="IconButtonVariant.Flat" class="icon-button" master-view-2-scope>
-                <span class="material-icons">
-                    message
-                </span>
-                <IgbRipple></IgbRipple>
-            </IgbIconButton>
-            <IgbAvatar Src="assets/Avatar6.png" Shape="AvatarShape.Circle" @onclick="@(e => _dropdown3!.ToggleAsync(_avatar))" @ref="_avatar" class="avatar" master-view-2-scope></IgbAvatar>
-            <IgbDropdown @ref="_dropdown3" class="dropdown" master-view-2-scope>
-                <IgbDropdownItem Disabled="true">
-                    Edit Profile
-                </IgbDropdownItem>
-                <IgbDropdownItem Disabled="true">
-                    About
-                </IgbDropdownItem>
-                <IgbDropdownItem>
-                    Logout
-                </IgbDropdownItem>
-            </IgbDropdown>
-        </div>
-    </div>
-    <div class="row-layout group_2">
-        <IgbNavDrawer Position="NavDrawerPosition.Relative" class="nav-drawer" master-view-2-scope>
-            <div slot="mini">
-                <IgbNavDrawerItem>
-                    <span slot="icon">
-                        <span class="material-icons icon">
-                            assignment_turned_in
-                        </span>
-                        <IgbRipple></IgbRipple>
-                    </span>
-                </IgbNavDrawerItem>
-                <IgbNavDrawerItem>
-                    <span slot="icon">
-                        <span class="material-icons icon">
-                            dashboard
-                        </span>
-                        <IgbRipple></IgbRipple>
-                    </span>
-                </IgbNavDrawerItem>
-                <IgbNavDrawerItem>
-                    <span slot="icon">
-                        <span class="material-icons icon">
-                            vertical_split
-                        </span>
-                        <IgbRipple></IgbRipple>
-                    </span>
-                </IgbNavDrawerItem>
-                <IgbNavDrawerItem>
-                    <span slot="icon">
-                        <span class="material-icons icon">
-                            bar_chart
-                        </span>
-                        <IgbRipple></IgbRipple>
-                    </span>
-                </IgbNavDrawerItem>
-                <IgbNavDrawerItem>
-                    <span slot="icon">
-                        <span class="material-icons icon">
-                            folder
-                        </span>
-                        <IgbRipple></IgbRipple>
-                    </span>
-                </IgbNavDrawerItem>
-            </div>
-            <IgbNavDrawerItem>
-                <span slot="icon">
-                    <span class="material-icons icon">
-                        account_circle
-                    </span>
-                    <IgbRipple></IgbRipple>
-                </span>
-                <div slot="content">Title goes here</div>
-            </IgbNavDrawerItem>
-        </IgbNavDrawer>
-        <div class="view-container">
-            @Body
-        </div>
-    </div>
-</div>
-
-@code {
-    private IgbButton? _button3;
-    private IgbDropdown? _projectDropdown;
-    private IgbAvatar? _avatar;
-    private IgbDropdown? _dropdown3;
-}
