+++ TestBugs in Samples/Pages/TestCardsOpacity.cs
@@ -1,10 +1,10 @@
 using Bunit;
 using Microsoft.Extensions.DependencyInjection;
-using Bugs_in_Samples.Pages.MasterView;
+using Bugs_in_Samples.Pages;
 
 namespace TestBugs_in_Samples
 {
-	public class TestButtons
+	public class TestCardsOpacity
 	{
 		[Fact]
 		public void ViewIsCreated()
@@ -12,42 +12,35 @@ namespace TestBugs_in_Samples
 			using var ctx = new TestContext();
 			ctx.JSInterop.Mode = JSRuntimeMode.Loose;
 			ctx.Services.AddIgniteUIBlazor(
+				typeof(IgniteUI.Blazor.Controls.IgbCardModule),
 				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
 				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
 				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
 				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
-				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
 				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
-				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
-				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
-				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
-				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
-				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
-				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
-				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
-				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
-				typeof(IgniteUI.Blazor.Controls.IgbDropdownModule),
-				typeof(IgniteUI.Blazor.Controls.IgbDropdownItemModule),
-				typeof(IgniteUI.Blazor.Controls.IgbListModule),
-				typeof(IgniteUI.Blazor.Controls.IgbAvatarModule),
-				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
 				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
 				typeof(IgniteUI.Blazor.Controls.IgbCardModule),
 				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
 				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
 				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
 				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbCardModule),
 				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
 				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
 				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
 				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
-				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
 				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
-				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
 				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
-				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
-				typeof(IgniteUI.Blazor.Controls.IgbRippleModule));
-			var componentUnderTest = ctx.RenderComponent<Buttons>();
+				typeof(IgniteUI.Blazor.Controls.IgbBadgeModule));
+			var componentUnderTest = ctx.RenderComponent<CardsOpacity>();
 			Assert.NotNull(componentUnderTest);
 		}
 	}
