+++ TestBugs in Samples/Pages/TestDialog.cs
@@ -1,10 +1,10 @@
 using Bunit;
 using Microsoft.Extensions.DependencyInjection;
-using Bugs_in_Samples.Pages.MasterView;
+using Bugs_in_Samples.Pages;
 
 namespace TestBugs_in_Samples
 {
-	public class TestCardActions
+	public class TestDialog
 	{
 		[Fact]
 		public void ViewIsCreated()
@@ -12,6 +12,10 @@ namespace TestBugs_in_Samples
 			using var ctx = new TestContext();
 			ctx.JSInterop.Mode = JSRuntimeMode.Loose;
 			ctx.Services.AddIgniteUIBlazor(
+				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
 				typeof(IgniteUI.Blazor.Controls.IgbCardModule),
 				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
 				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
@@ -34,8 +38,9 @@ namespace TestBugs_in_Samples
 				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
 				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
 				typeof(IgniteUI.Blazor.Controls.IgbBadgeModule),
+				typeof(IgniteUI.Blazor.Controls.IgbBadgeModule),
 				typeof(IgniteUI.Blazor.Controls.IgbBadgeModule));
-			var componentUnderTest = ctx.RenderComponent<CardActions>();
+			var componentUnderTest = ctx.RenderComponent<Dialog>();
 			Assert.NotNull(componentUnderTest);
 		}
 	}
