+++ TestBugs in Samples/Pages/TestNavigationBar.cs
@@ -1,0 +1,83 @@
+using Bunit;
+using Microsoft.Extensions.DependencyInjection;
+using Bugs_in_Samples.Pages;
+
+namespace TestBugs_in_Samples
+{
+	public class TestNavigationBar
+	{
+		[Fact]
+		public void ViewIsCreated()
+		{
+			using var ctx = new TestContext();
+			ctx.JSInterop.Mode = JSRuntimeMode.Loose;
+			ctx.Services.AddIgniteUIBlazor(
+				typeof(IgniteUI.Blazor.Controls.IgbNavbarModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbNavbarModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbNavbarModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbAvatarModule),
+				typeof(IgniteUI.Blazor.Controls.IgbNavbarModule),
+				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbAvatarModule),
+				typeof(IgniteUI.Blazor.Controls.IgbNavbarModule),
+				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbIconButtonModule),
+				typeof(IgniteUI.Blazor.Controls.IgbRippleModule),
+				typeof(IgniteUI.Blazor.Controls.IgbAvatarModule));
+			var componentUnderTest = ctx.RenderComponent<NavigationBar>();
+			Assert.NotNull(componentUnderTest);
+		}
+	}
+}
